`onContentLoaded()` 버그 수정
